requirements.txt
anyio==3.7.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.1
Flask==2.2.5
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dotenv==0.21.1
python-telegram-bot==20.3
requests==2.31.0
sniffio==1.3.0
typing-extensions==4.6.2
urllib3==2.0.2
Werkzeug==2.2.3
zipp==3.15.0
